shared/auth/pyproject.toml
[project]
name = "monorepo-shared-auth"
version = "0.1.0"
description = "Shared auth helpers for internal consumers"
requires-python = ">=3.14"
dependencies = [
    "PyJWT>=2.10.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/company_shared"]
